+++ R/RcppExports.R
@@ -1799,15 +1799,15 @@ initCochardNetworks <- function(x) {
 #' @param network A hydraulic network element of the list returned by \code{initCochardNetworks}
 #' @param dt Smallest time step (seconds)
 #' @param opt Option flag vector
-#' @param cavitationRecoveryStem,cavitationRecoveryLeaves A string indicating how refilling of embolized conduits is done:
+#' @param stemCavitationRecovery,leafCavitationRecovery A string indicating how refilling of embolized conduits is done:
 #'           \itemize{
 #'             \item{"none" - no refilling.}
 #'             \item{"annual" - every first day of the year.}
 #'             \item{"rate" - following a rate of new sapwood formation.}
 #'             \item{"total" - instantaneous complete refilling.}
 #'           }
-semi_implicit_integration <- function(network, dt, opt, cavitationRecoveryStem = "annual", cavitationRecoveryLeaves = "total") {
-    invisible(.Call(`_medfate_semi_implicit_integration`, network, dt, opt, cavitationRecoveryStem, cavitationRecoveryLeaves))
+semi_implicit_integration <- function(network, dt, opt, stemCavitationRecovery = "annual", leafCavitationRecovery = "total") {
+    invisible(.Call(`_medfate_semi_implicit_integration`, network, dt, opt, stemCavitationRecovery, leafCavitationRecovery))
 }
 
 #' @rdname hydraulics_supplyfunctions