+++ R/transp_maximumTranspirationModel.R
@@ -118,8 +118,8 @@ transp_maximumTranspirationModel<-function(x, meteo, latitude, elevation, slope,
   xIni = x
   xIni$control$modifyInput = FALSE
   xIni$control$unlimitedSoilWater = TRUE
-  xIni$control$cavitationRecoveryStem = "total"
-  xIni$control$cavitationRecoveryLeaves = "total"
+  xIni$control$stemCavitationRecovery = "total"
+  xIni$control$leafCavitationRecovery = "total"
   xIni$control$verbose = FALSE
   Tmax = matrix(NA, nrow=ndays, ncol = nlai)  
   colnames(Tmax) = LAI_seq