+++ src/hydraulics.cpp
@@ -706,10 +706,14 @@ List E2psiAboveground(double E, double psiRootCrown,
   double leafd = hydraulicNetwork["leafd"];
   double PLCstem = hydraulicNetwork["PLCstem"];
   double PLCleaf = hydraulicNetwork["PLCleaf"];
+  bool stemCavitationEffects = hydraulicNetwork["stemCavitationEffects"];
+  bool leafCavitationEffects = hydraulicNetwork["leafCavitationEffects"];
     
-  double psiPLCStem =  apoplasticWaterPotential(std::max(0.0001, 1.0 - PLCstem), stemc, stemd);
+  double psiPLCStem =  0.0;
+  if(stemCavitationEffects) psiPLCStem = apoplasticWaterPotential(std::max(0.0001, 1.0 - PLCstem), stemc, stemd);
   double psiStem = E2psiXylem(E, psiRootCrown, kstemmax, stemc, stemd, psiPLCStem); //Apliquem la fatiga per cavitacio a la caiguda de potencial a la tija 
-  double psiPLCLeaf=  apoplasticWaterPotential(std::max(0.0001,1.0-PLCleaf), leafc, leafd);
+  double psiPLCLeaf= 0.0;
+  if(leafCavitationEffects) psiPLCLeaf = apoplasticWaterPotential(std::max(0.0001,1.0-PLCleaf), leafc, leafd);
   double psiLeaf = E2psiXylem(E, psiStem, kleafmax, leafc, leafd, psiPLCLeaf); 
   return(List::create( Named("E")=E, Named("psiStem") =psiStem,Named("psiLeaf") =psiLeaf));
 }