+++ src/inner_cochard.cpp
@@ -489,7 +489,7 @@ void calculateRhizoPsi(int c,
 //' @param network A hydraulic network element of the list returned by \code{initCochardNetworks}
 //' @param dt Smallest time step (seconds)
 //' @param opt Option flag vector
-//' @param cavitationRecoveryStem,cavitationRecoveryLeaves A string indicating how refilling of embolized conduits is done:
+//' @param stemCavitationRecovery,leafCavitationRecovery A string indicating how refilling of embolized conduits is done:
 //'           \itemize{
 //'             \item{"none" - no refilling.}
 //'             \item{"annual" - every first day of the year.}
@@ -498,7 +498,7 @@ void calculateRhizoPsi(int c,
 //'           }
 // [[Rcpp::export("semi_implicit_integration")]]
 void semi_implicit_integration(List network, double dt, NumericVector opt, 
-                               String cavitationRecoveryStem = "annual", String cavitationRecoveryLeaves = "total") {
+                               String stemCavitationRecovery = "annual", String leafCavitationRecovery = "total") {
   
   List params = as<Rcpp::List>(network["params"]);
   NumericVector PsiSoil = network["PsiSoil"];
@@ -645,7 +645,7 @@ void semi_implicit_integration(List network, double dt, NumericVector opt,
   //# Cavitation
   psirefL = network["Psi_LApo"];  //# the reference is at current time step for other modes  (implicit, explicit)
   psirefS = network["Psi_SApo"];  //# The reference is at current time step for other modes (implicit, explicit)
-  if(cavitationRecoveryStem!="total") {
+  if(stemCavitationRecovery!="total") {
     if (psirefS < Psi_SApo_cav) {
       network["Psi_SApo_cav"] = psirefS;
       network["PLC_Stem"] = PLC(psirefS, VCstem_slope, VCstem_P50);
@@ -654,7 +654,7 @@ void semi_implicit_integration(List network, double dt, NumericVector opt,
     network["Psi_SApo_cav"] = psirefS;
     network["PLC_Stem"] = PLC(psirefS, VCstem_slope, VCstem_P50);
   }
-  if(cavitationRecoveryLeaves!="total") {
+  if(leafCavitationRecovery!="total") {
     if(psirefL < Psi_LApo_cav) {
       network["Psi_LApo_cav"] = psirefL;
       network["PLC_Leaf"] = PLC(psirefL, VCleaf_slope, VCleaf_P50);
@@ -684,8 +684,8 @@ void innerCochard(List x, List input, List output, int n, double tstep,
   // Extract control variables
   List control = x["control"];
   String soilFunctions = control["soilFunctions"];
-  String cavitationRecoveryStem = control["cavitationRecoveryStem"];
-  String cavitationRecoveryLeaves = control["cavitationRecoveryLeaves"];
+  String stemCavitationRecovery = control["stemCavitationRecovery"];
+  String leafCavitationRecovery = control["leafCavitationRecovery"];
   String rhizosphereOverlap = control["rhizosphereOverlap"];
   bool plantWaterPools = (rhizosphereOverlap!="total");
   bool plantCapacitance = control["plantCapacitance"];
@@ -1015,7 +1015,7 @@ void innerCochard(List x, List input, List output, int n, double tstep,
           network_n["Einst_SH"] = Elim_SH + Emin_L_SH; //For shade photosynthesis/transpiration
           
           //Effects on water potentials and flows
-          semi_implicit_integration(network_n, dt, opt, cavitationRecoveryStem, cavitationRecoveryLeaves);
+          semi_implicit_integration(network_n, dt, opt, stemCavitationRecovery, leafCavitationRecovery);
           update_conductances(network_n);
           update_capacitances(network_n);